split up bootstrap/ into multiple implementation/test pairs
