+++ bootstrap/BUILD
@@ -2,7 +2,12 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["bootstrap.go"],
+    srcs = [
+        "args.go",
+        "main.go",
+        "paths.go",
+        "repos.go",
+    ],
     visibility = ["//visibility:private"],
     deps = ["//vendor/github.com/spf13/pflag:go_default_library"],
 )
@@ -29,6 +34,9 @@ filegroup(
 
 go_test(
     name = "go_default_test",
-    srcs = ["bootstrap_test.go"],
+    srcs = [
+        "args_test.go",
+        "repos_test.go",
+    ],
     library = ":go_default_library",
 )
